# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TRIXHUB Affiliation Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface RegisterBody {
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  country: string;
  password: string;
  /** @nullable */
  referralCode?: string | null;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface ActivateBody {
  paymentMethod: string;
  paymentReference: string;
  currency?: string;
}

export interface UpdateProfileBody {
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  country?: string | null;
}

export interface UpdateCurrencyBody {
  currency: string;
}

export interface User {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  country: string;
  isActivated: boolean;
  referralCode: string;
  /** @nullable */
  referredByCode?: string | null;
  preferredCurrency: string;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface Dashboard {
  totalBalance: number;
  referralBalance: number;
  taskBalance: number;
  withdrawnAmount: number;
  spentAmount: number;
  inactiveBalance: number;
  totalReferrals: number;
  activeReferrals: number;
  inactiveReferrals: number;
  level1Count: number;
  level2Count: number;
  level3Count: number;
  pendingWithdrawals: number;
  currency: string;
  exchangeRate: number;
}

export interface ReferralUser {
  id: number;
  firstName: string;
  lastName: string;
  country: string;
  isActivated: boolean;
  joinedAt: string;
  level: number;
}

export interface TeamResponse {
  total: number;
  active: number;
  inactive: number;
  members: ReferralUser[];
}

export interface ReferralLevelResponse {
  level: number;
  commission: number;
  total: number;
  active: number;
  inactive: number;
  members: ReferralUser[];
}

export interface BalanceSummary {
  totalBalance: number;
  referralBalance: number;
  taskBalance: number;
  inactiveBalance: number;
  withdrawnAmount: number;
  spentAmount: number;
  minimumWithdrawal: number;
  currency: string;
  exchangeRate: number;
}

export interface WithdrawalBody {
  amount: number;
  method: string;
  accountNumber: string;
  accountName: string;
}

export interface Withdrawal {
  id: number;
  amount: number;
  method: string;
  accountNumber: string;
  accountName: string;
  status: string;
  requestedAt: string;
  /** @nullable */
  processedAt?: string | null;
}

export interface Task {
  id: number;
  title: string;
  description: string;
  reward: number;
  type: string;
  isCompleted: boolean;
  /** @nullable */
  completedAt?: string | null;
  /** @nullable */
  url?: string | null;
}

export interface PlatformConfig {
  activationFee: number;
  level1Commission: number;
  level2Commission: number;
  level3Commission: number;
  minimumWithdrawal: number;
  currency: string;
}

export type CurrencyRatesRates = { [key: string]: number };

export interface CurrencyRates {
  baseCurrency: string;
  rates: CurrencyRatesRates;
  updatedAt: string;
}

export interface ActivityItem {
  id: number;
  type: string;
  message: string;
  /** @nullable */
  amount?: number | null;
  createdAt: string;
}