# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TRIXHUB Affiliation Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/activate:
    post:
      operationId: activateAccount
      tags: [auth]
      summary: Activate user account (3600 FCFA)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivateBody"
      responses:
        "200":
          description: Account activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /referrals/team:
    get:
      operationId: getTeam
      tags: [referrals]
      summary: Get referral team
      responses:
        "200":
          description: Team list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /referrals/level/{level}:
    get:
      operationId: getReferralsByLevel
      tags: [referrals]
      summary: Get referrals by level
      parameters:
        - name: level
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Referrals at level
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferralLevelResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    RegisterBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        country:
          type: string
        password:
          type: string
        referralCode:
          type: ["string", "null"]
      required: [firstName, lastName, email, phone, country, password]
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    ActivateBody:
      type: object
      properties:
        paymentMethod:
          type: string
        paymentReference:
          type: string
        currency:
          type: string
      required: [paymentMethod, paymentReference]
...
    User:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        country:
          type: string
        isActivated:
          type: boolean
        referralCode:
          type: string
        referredByCode:
          type: ["string", "null"]
        preferredCurrency:
          type: string
        createdAt:
          type: string
      required: [id, firstName, lastName, email, phone, country, isActivated, referralCode, preferredCurrency, createdAt]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]
...
    ReferralUser:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        country:
          type: string
        isActivated:
          type: boolean
        joinedAt:
          type: string
        level:
          type: integer
      required: [id, firstName, lastName, country, isActivated, joinedAt, level]
    TeamResponse:
      type: object
      properties:
        total:
          type: integer
        active:
          type: integer
        inactive:
          type: integer
        members:
          type: array
          items:
            $ref: "#/components/schemas/ReferralUser"
      required: [total, active, inactive, members]
    ReferralLevelResponse:
      type: object
      properties:
        level:
          type: integer
        commission:
          type: number
        total:
          type: integer
        active:
          type: integer
        inactive:
          type: integer
        members:
          type: array
          items:
            $ref: "#/components/schemas/ReferralUser"
      required: [level, commission, total, active, inactive, members]
...