# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TRIXHUB Affiliation Platform API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Register a new user
 */
export const RegisterBody = zod.object({
  firstName: zod.string(),
  lastName: zod.string(),
  email: zod.string(),
  phone: zod.string(),
  country: zod.string(),
  password: zod.string(),
  referralCode: zod.string().nullish(),
});

/**
 * @summary Login
 */
export const LoginBody = zod.object({
  email: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  user: zod.object({
    id: zod.number(),
    firstName: zod.string(),
    lastName: zod.string(),
    email: zod.string(),
    phone: zod.string(),
    country: zod.string(),
    isActivated: zod.boolean(),
    referralCode: zod.string(),
    referredByCode: zod.string().nullish(),
    preferredCurrency: zod.string(),
    createdAt: zod.string(),
  }),
  token: zod.string(),
});

/**
 * @summary Logout
 */
export const LogoutResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().optional(),
});

/**
 * @summary Get current user
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  firstName: zod.string(),
  lastName: zod.string(),
  email: zod.string(),
  phone: zod.string(),
  country: zod.string(),
  isActivated: zod.boolean(),
  referralCode: zod.string(),
  referredByCode: zod.string().nullish(),
  preferredCurrency: zod.string(),
  createdAt: zod.string(),
});

/**
 * @summary Activate user account (3600 FCFA)
 */
export const ActivateAccountBody = zod.object({
  paymentMethod: zod.string(),
  paymentReference: zod.string(),
  currency: zod.string().optional(),
});

export const ActivateAccountResponse = zod.object({
  user: zod.object({
    id: zod.number(),
    firstName: zod.string(),
    lastName: zod.string(),
    email: zod.string(),
    phone: zod.string(),
    country: zod.string(),
    isActivated: zod.boolean(),
    referralCode: zod.string(),
    referredByCode: zod.string().nullish(),
    preferredCurrency: zod.string(),
    createdAt: zod.string(),
  }),
  token: zod.string(),
});

/**
 * @summary Update profile
 */
export const UpdateProfileBody = zod.object({
  firstName: zod.string().nullish(),
  lastName: zod.string().nullish(),
  phone: zod.string().nullish(),
  country: zod.string().nullish(),
});

export const UpdateProfileResponse = zod.object({
  id: zod.number(),
  firstName: zod.string(),
  lastName: zod.string(),
  email: zod.string(),
  phone: zod.string(),
  country: zod.string(),
  isActivated: zod.boolean(),
  referralCode: zod.string(),
  referredByCode: zod.string().nullish(),
  preferredCurrency: zod.string(),
  createdAt: zod.string(),
});

/**
 * @summary Get dashboard summary
 */
export const GetDashboardResponse = zod.object({
  totalBalance: zod.number(),
  referralBalance: zod.number(),
  taskBalance: zod.number(),
  withdrawnAmount: zod.number(),
  spentAmount: zod.number(),
  inactiveBalance: zod.number(),
  totalReferrals: zod.number(),
  activeReferrals: zod.number(),
  inactiveReferrals: zod.number(),
  level1Count: zod.number(),
  level2Count: zod.number(),
  level3Count: zod.number(),
  pendingWithdrawals: zod.number(),
  currency: zod.string(),
  exchangeRate: zod.number(),
});

/**
 * @summary Get referral team
 */
export const GetTeamResponse = zod.object({
  total: zod.number(),
  active: zod.number(),
  inactive: zod.number(),
  members: zod.array(
    zod.object({
      id: zod.number(),
      firstName: zod.string(),
      lastName: zod.string(),
      country: zod.string(),
      isActivated: zod.boolean(),
      joinedAt: zod.string(),
      level: zod.number(),
    }),
  ),
});

/**
 * @summary Get referrals by level
 */
export const GetReferralsByLevelParams = zod.object({
  level: zod.coerce.number(),
});

export const GetReferralsByLevelResponse = zod.object({
  level: zod.number(),
  commission: zod.number(),
  total: zod.number(),
  active: zod.number(),
  inactive: zod.number(),
  members: zod.array(
    zod.object({
      id: zod.number(),
      firstName: zod.string(),
      lastName: zod.string(),
      country: zod.string(),
      isActivated: zod.boolean(),
      joinedAt: zod.string(),
      level: zod.number(),
    }),
  ),
});

/**
 * @summary Get all balances
 */
export const GetBalancesResponse = zod.object({
  totalBalance: zod.number(),
  referralBalance: zod.number(),
  taskBalance: zod.number(),
  inactiveBalance: zod.number(),
  withdrawnAmount: zod.number(),
  spentAmount: zod.number(),
  minimumWithdrawal: zod.number(),
  currency: zod.string(),
  exchangeRate: zod.number(),
});

/**
 * @summary List withdrawal requests
 */
export const ListWithdrawalsResponseItem = zod.object({
  id: zod.number(),
  amount: zod.number(),
  method: zod.string(),
  accountNumber: zod.string(),
  accountName: zod.string(),
  status: zod.string(),
  requestedAt: zod.string(),
  processedAt: zod.string().nullish(),
});
export const ListWithdrawalsResponse = zod.array(ListWithdrawalsResponseItem);

/**
 * @summary Request withdrawal
 */
export const RequestWithdrawalBody = zod.object({
  amount: zod.number(),
  method: zod.string(),
  accountNumber: zod.string(),
  accountName: zod.string(),
});

/**
 * @summary List tasks and missions
 */
export const ListTasksResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  reward: zod.number(),
  type: zod.string(),
  isCompleted: zod.boolean(),
  completedAt: zod.string().nullish(),
  url: zod.string().nullish(),
});
export const ListTasksResponse = zod.array(ListTasksResponseItem);

/**
 * @summary Mark task as complete
 */
export const CompleteTaskParams = zod.object({
  id: zod.coerce.number(),
});

export const CompleteTaskResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  reward: zod.number(),
  type: zod.string(),
  isCompleted: zod.boolean(),
  completedAt: zod.string().nullish(),
  url: zod.string().nullish(),
});

/**
 * @summary Get platform config and commission rates
 */
export const GetPlatformConfigResponse = zod.object({
  activationFee: zod.number(),
  level1Commission: zod.number(),
  level2Commission: zod.number(),
  level3Commission: zod.number(),
  minimumWithdrawal: zod.number(),
  currency: zod.string(),
});

/**
 * @summary Get currency exchange rates vs FCFA
 */
export const GetCurrencyRatesResponse = zod.object({
  baseCurrency: zod.string(),
  rates: zod.record(zod.string(), zod.number()),
  updatedAt: zod.string(),
});

/**
 * @summary Update preferred currency
 */
export const UpdatePreferredCurrencyBody = zod.object({
  currency: zod.string(),
});

export const UpdatePreferredCurrencyResponse = zod.object({
  id: zod.number(),
  firstName: zod.string(),
  lastName: zod.string(),
  email: zod.string(),
  phone: zod.string(),
  country: zod.string(),
  isActivated: zod.boolean(),
  referralCode: zod.string(),
  referredByCode: zod.string().nullish(),
  preferredCurrency: zod.string(),
  createdAt: zod.string(),
});

/**
 * @summary Get recent referral activity feed
 */
export const GetReferralActivityResponseItem = zod.object({
  id: zod.number(),
  type: zod.string(),
  message: zod.string(),
  amount: zod.number().nullish(),
  createdAt: zod.string(),
});
export const GetReferralActivityResponse = zod.array(
  GetReferralActivityResponseItem,
);